+++ server/routes/items.js
@@ -27,7 +27,7 @@ exports.getItemsByModuleUuid = async (req, res) => {
       return res.status(404).send()
     }
 
-    const items = await Item.find({ module: module._id }).populate('module')
+    const items = await Item.find({ module: module._id }).populate('module').sort('order')
     res.send(items)
   } catch (e) {
     res.send(400).send(e)
@@ -96,6 +96,13 @@ exports.deleteItem = async (req, res) => {
     if (!item) {
       return res.status(404).send()
     }
+
+    const module = await Module.findByIdAndUpdate(item.module, { $pull: { items: item._id } })
+
+    if (!module) {
+      return res.status(404).send()
+    }
+
     res.send({ item: item.toPublic() })
   } catch (e) {
     res.status(404).send()