+++ server/server.js
@@ -1,10 +1,19 @@
 require('./config/config')
 
 const express = require('express')
+const morgan = require('morgan')
 const bodyParser = require('body-parser')
 const { mongoose } = require('./db/mongoose')
+const cors = require('cors')
 
 const app = express()
+
+if (process.env.NODE_ENV !== 'test') {
+  app.use(morgan('combined'))
+}
+
+app.use(cors())
+
 const port = process.env.PORT
 const routes = require('./routes')